+++ libvb/vb/Array.h
@@ -9,14 +9,14 @@ namespace vb {
 	template <typename T> class Array_iterator;
 
 	template <typename T> class Array { public:
-		Array (int w, int h, T d) :  ww(w), hh(h), data (w*h,d) {};
-		Array (int w=0, int h=0) :  ww(w), hh(h), data (w*h) {};
+		Array (long w, long h, T d) :  ww(w), hh(h), data (w*h,d) {};
+		Array (long w=0, long h=0) :  ww(w), hh(h), data (w*h) {};
 		Array (const std::vector<std::vector<T>> & l) : ww(l.size()), hh(l[0].size()), data(ww*hh) {
-			for (int i=0; i<ww; ++i) for (int j=0; j<hh; ++j) put(coo(i,j),l[i][j]);
+			for (long i=0; i<ww; ++i) for (long j=0; j<hh; ++j) put(coo(i,j),l[i][j]);
 		}
 
-		void resize (int w, int h) { ww=w; hh=h; data.resize(ww*hh); }
-		void resize (int w, int h, T t) { ww=w; hh=h; data.resize(ww*hh,t); }
+		void resize (long w, long h) { ww=w; hh=h; data.resize(ww*hh); }
+		void resize (long w, long h, T t) { ww=w; hh=h; data.resize(ww*hh,t); }
 
 		T &      	at	(coo z)      	{ return data[z.x + ww*z.y]; }
 		T const &	at	(coo z) const	{ return data[z.x + ww*z.y]; }
@@ -24,20 +24,20 @@ namespace vb {
 		T &      	operator[]	(coo z)      	{ return at(z); }
 		T const &	operator[]	(coo z) const	{ return at(z); }
 
-		T &      	atp	(coo z)      	{ int x = pmod (z.x,ww), y = pmod (z.y,hh); return at(coo(x,y)); }
-		T const &	atp	(coo z) const	{ int x = pmod (z.x,ww), y = pmod (z.y,hh); return at(coo(x,y)); }
+		T &      	atp	(coo z)      	{ long x = pmod (z.x,ww), y = pmod (z.y,hh); return at(coo(x,y)); }
+		T const &	atp	(coo z) const	{ long x = pmod (z.x,ww), y = pmod (z.y,hh); return at(coo(x,y)); }
 
 		void	put 	(coo z, T const & c)	{ at(z) = c; }
 		void	putp	(coo z, T const & c)	{ atp(z) = c; }
 
-		bool	contains	(coo z, int b=0)	const	{ return (z.x>=b) && (z.y>=b) && (z.x<ww-b) && (z.y<hh-b); }
-		coo 	rand    	(int b=0)       	const	{ return coo (b+prng.uniform_int(ww-2*b), b+prng.uniform_int(hh-2*b)); }
+		bool	contains	(coo z, long b=0)	const	{ return (z.x>=b) && (z.y>=b) && (z.x<ww-b) && (z.y<hh-b); }
+		coo 	rand    	(long b=0)       	const	{ return coo (b+prng.uniform_int(ww-2*b), b+prng.uniform_int(hh-2*b)); }
 
 		using iterator = Array_iterator<T>;
 		iterator	begin	()	{ return iterator (*this, coo(0,0)); 	}
 		iterator	end  	()	{ return iterator (*this, coo(0,hh));	}
 
-		int ww,hh;
+		long ww,hh;
 
 	private:
 		std::vector<T> data;
@@ -52,6 +52,6 @@ namespace vb {
 	};
 
 	template <typename T> Stream<coo> coos (const Array<T> & A) { return Stream<coo> ([&A](Sink<coo> & yield) {
-		for (int y=0; y<A.hh; ++y) for (int x=0; x<A.ww; ++x) yield(coo(x,y));
+		for (long y=0; y<A.hh; ++y) for (long x=0; x<A.ww; ++x) yield(coo(x,y));
 	});}
 }
